成果pageリファクタリング
